[FO-2389] flytter dialogvisningen inn i ny dialogoversikt reanamer gamel dialogoversikt til dialogliste
