+++ src/view/App.tsx
@@ -1,13 +1,10 @@
 import React from 'react';
-import DialogOversikt from './dialogoversikt/DialogOversikt';
 import AlertStripeContainer from './alertstriper/AlertStripeContainer';
 import { AppBanner } from './banner/AppBanner';
-import AktivitetContainer from './aktivitet/AktivitetContainer';
-import Routes from './Routes';
 import './App.less';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from './Provider';
-import HidenIfBrukerAldriUnderOppfolging from './HidenIfBrukerAldriUnderOppfolging';
+import DialogOversikt from './DialogOversikt';
 
 function App() {
     return (
@@ -16,11 +13,7 @@ function App() {
                 <Provider>
                     <AppBanner />
                     <AlertStripeContainer />
-                    <HidenIfBrukerAldriUnderOppfolging>
-                        <DialogOversikt />
-                        <Routes />
-                        <AktivitetContainer />
-                    </HidenIfBrukerAldriUnderOppfolging>
+                    <DialogOversikt />
                 </Provider>
             </div>
         </Router>
