+++ src/view/dialogliste/DialogPreview.less
