+++ src/view/dialogliste/DialogPreview.tsx
